dom/xml/test/test_bug1666613.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Bug 1666613</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1666613">Mozilla Bug 691215</a>
<p id="display"></p>
<iframe id="target" src="file_bug1666613.xml" width="90%"></iframe>
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

const iframe = document.getElementById("target");

function checkDir(expected) {
  let parserError = iframe.contentDocument.documentElement;
  is(parserError.getAttribute("dir"), expected);
}

async function sanityTest() {
  iframe.onload = () => {
    checkDir("ltr");
    pseudoBidiTest();
  };
  iframe.contentWindow.location.reload();
}

async function pseudoBidiTest() {
  await SpecialPowers.pushPrefEnv({
    set: [["intl.l10n.pseudo", "bidi"]]
  });
  iframe.onload = () => {
    checkDir("rtl");
    SimpleTest.finish();
  };
  iframe.contentWindow.location.reload();
}

sanityTest();
</script>
</body>
</html>
